# docs/swagger.yaml
definitions:
  api.Error:
    properties:
      error:
        type: string
    type: object
  api.Job:
    properties:
      args:
        items:
          type: string
        type: array
      customer_id:
        type: string
      end_time:
        $ref: '#/definitions/timestamppb.Timestamp'
      error:
        type: string
      id:
        type: string
      input_id:
        type: string
      output_id:
        type: string
      start_time:
        $ref: '#/definitions/timestamppb.Timestamp'
      status:
        type: string
      task_type:
        type: string
    type: object
  rototiller.Storage:
    properties:
      create_time:
        $ref: '#/definitions/timestamppb.Timestamp'
      customer_id:
        type: string
      id:
        type: string
      last_used:
        $ref: '#/definitions/timestamppb.Timestamp'
      name:
        type: string
      status:
        type: string
    type: object
  rototiller.Task:
    properties:
      kind:
        type: string
      params:
        items:
          type: string
        type: array
      type:
        type: string
    type: object
  timestamppb.Timestamp:
    properties:
      nanos:
        description: |-
          Non-negative fractions of a second at nanosecond resolution. Negative
          second values with fractions must still have non-negative nanos values
          that count forward in time. Must be from 0 to 999,999,999
          inclusive.
        type: integer
      seconds:
        description: |-
          Represents seconds of UTC time since Unix epoch
          1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to
          9999-12-31T23:59:59Z inclusive.
        type: integer
    type: object
info:
  contact:
    email: logsquaredn@gmail.com
    name: logsquaredn
    url: https://rototiller.logsquaredn.io/
  license:
    name: logsquaredn
paths:
...
  /api/v1/jobs:
    get:
      description: |-
        Get a list of jobs based on API Key

        &emsp; - API Key is required either as a query parameter or a header
      parameters:
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      - description: API Key query parameter
        in: query
        name: api-key
        type: string
      - description: Offset of jobs to return
        in: query
        name: offset
        type: integer
      - description: Limit of jobs to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.Job'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get a list of jobs
      tags:
      - Job
  /api/v1/jobs/{id}:
    get:
      description: |-
        Get the metadata of a job. This can be used as a way to track job status

        &emsp; - API Key is required either as a query parameter or a header
      parameters:
      - description: API Key query parameter
        in: query
        name: api-key
        type: string
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Job'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get a job
      tags:
      - Job
  /api/v1/jobs/{id}/storages/input:
    get:
      description: |-
        Get the metadata of a job's input

        &emsp; - API Key is required either as a query parameter or a header
      parameters:
      - description: API Key query parameter
        in: query
        name: api-key
        type: string
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rototiller.Storage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get a job's input
      tags:
      - Storage
  /api/v1/jobs/{id}/storages/input/content:
    get:
      description: |-
        Gets the content of a job's input

        &emsp; - API Key is required either as a query parameter or a header
      parameters:
      - description: Request results as a Zip or JSON. Default Zip
        in: header
        name: Accept
        type: string
      - description: API Key query parameter
        in: query
        name: api-key
        type: string
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      - ' application/zip'
      responses:
        "200":
          description: ""
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get a job's input content
      tags:
      - Content
  /api/v1/jobs/{id}/storages/output:
    get:
      description: |-
        Get the metadata of a job's output

        &emsp; - API Key is required either as a query parameter or a header
      parameters:
      - description: API Key query parameter
        in: query
        name: api-key
        type: string
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rototiller.Storage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get a job's output
      tags:
      - Storage
  /api/v1/jobs/{id}/storages/output/content:
    get:
      description: |-
        Gets the content of a job's output

        &emsp; - API Key is required either as a query parameter or a header
      parameters:
      - description: Request results as a Zip or JSON. Default Zip
        in: header
        name: Accept
        type: string
      - description: API Key query parameter
        in: query
        name: api-key
        type: string
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      - ' application/zip'
      responses:
        "200":
          description: ""
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get a job's output content
      tags:
      - Content
  /api/v1/jobs/{id}/task:
    get:
      description: |-
        Get the metadata of a job's task

        &emsp; - API Key is required either as a query parameter or a header
      parameters:
      - description: API Key query parameter
        in: query
        name: api-key
        type: string
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rototiller.Task'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get a job's task
      tags:
      - Task
  /api/v1/jobs/buffer:
    post:
      consumes:
      - application/json
      - ' application/zip'
      description: |-
        <b><u>Create a buffer job</u></b>
        &emsp; - Buffers every geometry by the given distance

        &emsp; - For extra info: https://gdal.org/api/vector_c_api.html#_CPPv412OGR_G_Buffer12OGRGeometryHdi
        &emsp; - API Key is required either as a query parameter or a header
        &emsp; - Pass the geospatial data to be processed in the request body OR
        &emsp; - Pass the ID of an existing dataset with an empty request body
        &emsp; - This task accepts a ZIP containing a shapefile or GeoJSON input
        &emsp; - This task will automatically generate both GeoJSON and ZIP (shapfile) output
      parameters:
      - description: API Key query parameter
        in: query
        name: api-key
        type: string
      - description: Required if passing geospatial data in request body
        in: header
        name: Content-Type
        type: string
      - description: API Key header
        in: header
        name: X-API-Key
        type: string
      - description: ID of existing dataset to use
        in: query
        name: input
        type: string
      - description: ID of existing job whose input dataset to use
        in: query
        name: input-of
        type: string
      - description: ID of existing job whose output dataset to use
        in: query
        name: output-of
        type: string
      - description: Buffer distance
        in: query
        name: buffer-distance
        required: true
        type: integer
      - description: Quadrant Segment count
        in: query
        name: quadrant-segment-count
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Job'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Create a buffer job
      tags:
      - Job
...
swagger: "2.0"